+++ src/Mu.Modelling/Feature.Semantics.cs
@@ -1,6 +1,8 @@
 ﻿namespace Mu.Modelling
 {
+    using System.Collections.Immutable;
     using Fluentify;
+    using MooVC.Syntax.CSharp;
     using Valuify;
     using Ignore = Valuify.IgnoreAttribute;
 
@@ -18,6 +20,8 @@
 
             [Ignore]
             public bool IsOutOfScope => this == OutOfScope;
+
+            public ImmutableArray<Qualification> Transforms { get; set; } = ImmutableArray<Qualification>.Empty;
         }
     }
 }