+++ CHANGELOG.md
@@ -13,4 +13,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Muify identifier implicit conversion generation.
 - Added Muify identifier comparability generation for IComparable and CompareTo.
 - Added Muify identifier comparability generation for greater-than and less-than operators.
+- Added Muify default feature transform generation.
 - Added Muify feature fact generation.